Renamed task file, clip I when error is small. Cleaned includes.
the output of swings and turns now clips when the error is within startI for turns and swings.  this helps the turns be a lot smoother.  this doesnt happen when the turn is smaller then startI

Removed util.hpp and moved it to Helper.hpp.  moved some enums to helper.hpp