#21 Added tracking wheel support
tick_per_inch now changes how it calculates depending on if trakcing wheels are used or not.  there are 3 drive constructors, one for internal encoders, one for tracking wheels in brain, and one for tracking wheels in an expander. left_sensor and right_sensor output motor/tracking wheel depending on the constructor used.