Added more tracker support #165
With this comes a new constructor that is unique to not overlap with "ratio" used previously, and removing "ratio" will create breaking changes.  This new constructor takes in a motor group and it extracts that out into the ez motor groups.  This currently doesn't work as there's a bug with `get_direction` that has an open issue on the PROS kernel.

If a left and right tracker are enabled, the main sensors for the drive switch over to being these tracking wheels.  If only one sensor is used, then this tracker is only used for odom.  Normal drive motions will use IMEs still, as that PID uses a left and a right PID.

Only one horizontal tracker will be used.

Angle is currently always calculated with IMU.  Until there's an easy way to tune track width, #45 will remain open.