Refactoring and cleaned example project
Limited the amount of decimals that can be printed to the brain screen.  Some new functions were added to handle this and they are now being used in other places in EZ-Template.

Tracker screen and xya screen are now combined.  Users can see track width on the brain screen now, and will only see active trackers.